+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 [metosin/ring-swagger "0.20.3-SNAPSHOT"] is available but we use "0.20.2"
 [prismatic/plumbing "0.4.3"] is available but we use "0.4.2"
 [peridot "0.4.0"] is available but we use "0.3.1"
+[compojure "1.3.4"] is available but we use "1.3.3"
+[lein-ring "0.9.4"] is available but we use "0.9.3"
 ```
 
 ## 0.20.1 (2.5.2015)